copy_report/requirements.txt
requests==2.32.3
httpx
dependency-injector==4.45.0
pydantic-settings==2.6.1
oracledb==2.5.1
openpyxl
msal
requests